% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clere-class.R
\docType{methods}
\name{clusters}
\alias{clusters}
\alias{clusters-methods}
\alias{clusters,Clere-method}
\title{clusters method}
\usage{
clusters(object, threshold = NULL, ...)
}
\arguments{
\item{object}{[Clere]: Output object from \code{\link{fitClere}}.}

\item{threshold}{[numeric]: A numerical \code{threshold > 0.5}. If
\code{threshold = NULL} then the each variable is assigned to the cluster
having the largest associated posterior probability.}

\item{...}{Additional arguments, not to be supplied in this version.}
}
\description{
This function makes returns the estimated clustering of variables.
}
\seealso{
Overview : \code{\link{clere-package}} \cr 
Classes : \code{\linkS4class{Clere}} \cr 
Methods : \code{\link{show}}, \code{\link{plot}}, \code{\link{clusters}}, \code{\link{predict}}, \code{\link{summary}} \cr 
Functions : \code{\link{fitClere}} 
Datasets : \code{\link{numExpRealData}}, \code{\link{numExpSimData}}
}
