\name{is.clere}

\alias{is.clere}
\alias{is.clere-methods}
\alias{is.clere,ANY-method}

\title{
Is an \code{\linkS4class{Clere}} object
}

\description{
\code{is.clere} returns \code{TRUE} if \code{x} is an \code{\linkS4class{Clere}} object and \code{FALSE} otherwise.
}

\usage{
is.clere(object)
}

\arguments{
    \item{object}{[any]: object to be tested.}
}

\seealso{
Overview : \code{\link{clere-package}} \cr
Classes : \code{\linkS4class{Clere}} \cr
Methods :  \code{\link{show}}, \code{\link{plot}}, \code{\link{ggPlot}}, \code{\link{clusters}}, \code{\link{predict}}, \code{\link{summary}} \cr
Functions : \code{\link{fit.clere}}
}

\author{Loic Yengo \email{loic.yengo@gmail.com}}

\keyword{is.clere}
\keyword{is}
\keyword{methods}
\keyword{method}

% \examples{}
