\name{summary}

\docType{methods}

\alias{summary}
\alias{summary-methods}
\alias{summary,Clere-method}

\title{
summary method
}

\description{
This function summarizes the output of function \code{\link{fit.clere}}.
}

\usage{
\S4method{summary}{Clere}(object, ...)
}

\arguments{
    \item{object}{[Clere]: Output object from \code{\link{fit.clere}}.}
    \item{...}{Additional arguments, not to be supplied in this version.}
}

\seealso{
Overview : \code{\link{clere-package}} \cr
Classes : \code{\linkS4class{Clere}} \cr
Methods :  \code{\link{show}}, \code{\link{plot}}, \code{\link{ggPlot}}, \code{\link{clusters}}, \code{\link{predict}}, \code{\link{summary}} \cr
Functions : \code{\link{fit.clere}}
}

\author{Loic Yengo \email{loic.yengo@gmail.com}}

\keyword{summary}
\keyword{methods}
\keyword{method}
\keyword{Clere}
\keyword{clere}

% \examples{}
