% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clere-class.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary-methods}
\alias{summary,Clere-method}
\title{summary method}
\usage{
\S4method{summary}{Clere}(object, ...)
}
\arguments{
\item{object}{[Clere]: Output object from \code{\link{fitClere}}.}

\item{...}{Additional arguments, not to be supplied in this version.}
}
\description{
This function summarizes the output of function \code{\link{fitClere}}.
}
\seealso{
Overview : \code{\link{clere-package}} \cr 
Classes : \code{\linkS4class{Clere}} \cr 
Methods : \code{\link{show}}, \code{\link{plot}}, \code{\link{clusters}}, \code{\link{predict}}, \code{\link{summary}} \cr 
Functions : \code{\link{fitClere}} 
Datasets : \code{\link{numExpRealData}}, \code{\link{numExpSimData}}
}
