% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clere-class.R
\docType{class}
\name{Clere-class}
\alias{Clere-class}
\alias{[,Clere-method}
\alias{[<-,Clere-method}
\alias{show,Clere-method}
\alias{sClere-class}
\alias{show,sClere-method}
\alias{[,sClere-method}
\alias{[<-,sClere-method}
\title{\code{\linkS4class{Clere}} class}
\description{
This class contains all the input parameters to run CLERE.
}
\details{
\describe{
\item{y}{\link{numeric}: The vector of observed responses.}
\item{x}{\link{matrix}: The matrix of predictors.}
\item{n}{\link{integer}: The sample size or the number of rows in matrix x.}
\item{p}{\link{integer}: The number of variables of the number of columns in matrix x.}
\item{g}{\link{integer}: The number or the maximum number of groups considered. Maximum number of groups stands when model selection is required.}
\item{nItMC}{\link{numeric}: Number of Gibbs iterations to generate the partitions.}
\item{nItEM}{\link{numeric}: Number of SEM/MCEM iterations.}
\item{nBurn}{\link{numeric}: Number of SEM iterations discarded before calculating the MLE which is averaged over SEM draws.}
\item{dp}{\link{numeric}: Number of iterations between sampled partitions when calculating the likelihood at the end of the run.}
\item{nsamp}{\link{numeric}: Number of sampled partitions for calculating the likelihood at the end of the run.}
\item{sparse}{\link{logical}: Should a \code{0} class be imposed to the model?}
\item{analysis}{\link{character}: Which analysis is to be performed. Values are \code{"fit"}, \code{"bic"}, \code{"aic"} and \code{"icl"}.}
\item{algorithm}{\link{character}: The algorithmto be chosen to fit the model. Either the SEM-Gibbs algorithm or the MCEM algorithm. The most efficient algorithm being the SEM-Gibbs approach. MCEM is not available for binary response.}
\item{initialized}{\link{logical}: Is set to TRUE when an initial partition and an initial vector of parameters is given by the user.} \item{maxit}{\link{numeric}: An EM algorithm is used inside the SEM to maximize the complete log-likelihood \code{p(y,Z|theta)}. \code{maxit} stands as the maximum number of EM iterations for the internal EM.}
\item{tol}{\link{numeric}: Maximum increased in complete log-likelihood for the internal EM (stopping criterion).}
\item{seed}{\link{integer}: An integer given as a seed for random number generation. If set to \code{NULL}, then a random seed is generated between \code{1} and \code{1000}.}
\item{b}{\link{numeric}: Vector of parameter b. Its size equals the number of group(s).}
\item{pi}{\link{numeric}: Vector of parameter pi. Its size equals the number of group(s).}
\item{sigma2}{\link{numeric}: Parameter sigma^2.}
\item{gamma2}{\link{numeric}: Parameter gamma^2.} itemintercept\link{numeric}: Parameter beta_0 (intercept).
\item{likelihood}{\link{numeric}: Approximated log-likelihood.}
\item{entropy}{\link{numeric}: Approximated entropy.}
\item{P}{\link{matrix}: A \link{\code{p x g}} matrix of posterior probability of membership to the groups. \code{P = E[Z|theta]}.}
\item{theta}{\link{matrix}: A \link{\code{nItEM x (2g+4)}} matrix containing values of the model parameters and complete data likelihood at each iteration of the SEM/MCEM algorithm}
\item{Bw}{\link{matrix}: A \link{\code{p x nsamp}} matrix which columns are samples from the posterior distribution of Beta (regression coefficients) given the data and the maximum likelihood estimates.}
\item{Zw}{\link{matrix}: A \link{\code{p x nsamp}} matrix which columns are samples from the posterior distribution of Z (groups membership indicators) given the data and the maximum likelihood estimates.}
\item{theta0}{\link{numeric}: A vector size \link{\code{2g+3}} containing initial guess of the model parameters. See example for function \code{\link{fitClere}.}}
\item{Z0}{\link{numeric}: A \link{\code{p x 1}} vector of integers taking values between 1 and \code{p} (number of variables).}
}
}
\section{Methods}{

\describe{
\item{object\link{"slotName"}:}{Get the value of the field \code{slotName}.}
\item{object\link{"slotName"}<-value:}{Set \code{value} to the field \code{slotName}.}
\item{show(object):}{Returns the formatted values of \code{\linkS4class{Clere}} object.}
\item{plot(x, ...):}{Graphical summary for MCEM/SEM-Gibbs estimation.}
\item{clusters(object, threshold = NULL, ...):}{Returns the estimated clustering of variables.}
\item{predict(object, newx, ...):}{Returns prediction using a fitted model and a new matrix of design.}
\item{summary(object, ...):}{summarizes the output of function \code{\link{fitClere}}.}
}
}

\seealso{
Overview : \code{\link{clere-package}} \cr Classes :
\code{\linkS4class{Clere}} \cr Methods : \code{\link{show}},
\code{\link{plot}}, \code{\link{clusters}}, \code{\link{predict}},
\code{\link{summary}} \cr Functions : \code{\link{fitClere}},
\code{\link{fitPacs}} Datasets : \code{\link{numExpRealData}},
\code{\link{numExpSimData}}
}
\keyword{Clere}
\keyword{class}
\keyword{method}
\keyword{methods}
