% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clere-class.R
\docType{methods}
\name{plot-methods}
\alias{plot-methods}
\alias{plot}
\alias{plot,Clere-method}
\alias{plot,Clere,ANY-method}
\title{plot method}
\usage{
\S4method{plot}{Clere}(x, y, ...)
}
\arguments{
\item{x}{[Clere]: Output object from \code{\link{fitClere}}.}

\item{y}{[any]: Unused parameter.}

\item{...}{Additional arguments, not to be supplied in this version.}
}
\description{
Graphical summary for MCEM/SEM-Gibbs estimation.  This function represents
the course of the model parameters in view of the iterations of the
estimation algorithms implemented in \code{\link{fitClere}}.
}
\seealso{
Overview : \code{\link{clere-package}} \cr 
Classes : \code{\linkS4class{Clere}}, \code{\linkS4class{Pacs}} \cr 
Methods : \code{\link{plot}}, \code{\link{clusters}}, \code{\link{predict}}, \code{\link{summary}} \cr 
Functions : \code{\link{fitClere}}, \code{\link{fitPacs}} 
Datasets : \code{\link{numExpRealData}}, \code{\link{numExpSimData}}, \code{\link{algoComp}}
}
